Reorganize tests to new structure
