+++ tests/CookieMonster/core/setup.js
