+++ tests/CookieMonster/game/golden-cookie.js
