+++ tests/CookieMonster/game/heavenly.js
