+++ tests/CookieMonster/helpers/browser.js
