+++ tests/CookieMonster/helpers/cache.js
