+++ tests/CookieMonster/helpers/emphasizers.js
