+++ tests/CookieMonster/helpers/math.js
