+++ tests/CookieMonster/helpers/time.js
