+++ tests/CookieMonster/interface/bottom-bar.js
