+++ tests/CookieMonster/interface/settings.js
