+++ tests/CookieMonster/interface/tooltips.js
